Upgraded chain length to a 64 bit integer so that chain lengths > 2.1 billion can be done. This addresses issue 540. Runs OK, accepts long chain lengths but not actually tested a run of > 2.1 billion.

+++ src/dr/inference/mcmc/MCMCOptions.java
@@ -34,7 +34,7 @@ package dr.inference.mcmc;
  */
 public class MCMCOptions {
 
-    private int chainLength;
+    private long chainLength;
     private int fullEvaluationCount = 2000;
     private int minOperatorCountForFullEvaluation = 1;
     private boolean coercion = true;
@@ -47,7 +47,7 @@ public class MCMCOptions {
     /**
      * @return the chain length of the MCMC analysis
      */
-    public final int getChainLength() {
+    public final long getChainLength() {
         return chainLength;
     }
 
@@ -72,7 +72,7 @@ public class MCMCOptions {
         this.temperature = temperature;
     }
 
-    public final void setChainLength(int length) {
+    public final void setChainLength(long length) {
         chainLength = length;
     }
